requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.4.26
certvalidator==0.11.1
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
endesive==2.18.5
fonttools==4.57.0
fpdf2==2.8.3
google-auth==2.39.0
google-genai==1.12.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
Markdown==3.8
oscrypto==1.3.0
paramiko==3.5.1
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
PyKCS11==1.5.17
PyNaCl==1.5.0
PyPDF2==3.0.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.0
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
